requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.128.8
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
packaging==26.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.4.2
starlette==0.49.3
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
mysql-connector-python